renames ModVariableNamesVisitor -> ModVariableSchemasVisitor
